Move *Identifier into src/keras_format (8/n)
