+++ src/constraints_test.ts
@@ -12,8 +12,9 @@
 
 import {serialization, Tensor1D, tensor1d} from '@tensorflow/tfjs-core';
 
-import {ConstraintIdentifier, deserializeConstraint, getConstraint, serializeConstraint} from './constraints';
+import {deserializeConstraint, getConstraint, serializeConstraint} from './constraints';
 import * as tfl from './index';
+import {ConstraintIdentifier} from './keras_format/constraint_config';
 import {describeMathCPU, expectNoLeakedTensors, expectTensorsClose} from './utils/test_utils';
