+++ src/keras_format/core_layers_config.ts
@@ -9,13 +9,10 @@
  */
 
 import {ActivationIdentifier} from '../activations';
-import {ConstraintIdentifier} from '../constraints';
-import {InitializerIdentifier} from '../initializers';
-import {RegularizerIdentifier} from '../regularizers';
 
-import {ConstraintConfig} from './constraint_config';
-import {InitializerConfig} from './initializer_config';
-import {RegularizerConfig} from './regularizer_config';
+import {ConstraintConfig, ConstraintIdentifier} from './constraint_config';
+import {InitializerConfig, InitializerIdentifier} from './initializer_config';
+import {RegularizerConfig, RegularizerIdentifier} from './regularizer_config';
 import {LayerPrimitiveArgs} from './topology_config';
 import {Shape} from './types';
