+++ src/layers/advanced_activations.ts
@@ -20,8 +20,9 @@ import {cast} from '../backend/tfjs_backend';
 import {Constraint, getConstraint, serializeConstraint} from '../constraints';
 import {InputSpec, Layer, LayerNonSerializableArgs} from '../engine/topology';
 import {NotImplementedError, ValueError} from '../errors';
-import {getInitializer, Initializer, InitializerIdentifier, serializeInitializer} from '../initializers';
+import {getInitializer, Initializer, serializeInitializer} from '../initializers';
 import {ELULayerPrimitiveArgs, LeakyReLULayerPrimitiveArgs, PReLULayerPrimitiveArgs, ReLULayerPrimitiveArgs, SoftmaxLayerPrimitiveArgs, ThresholdedReLULayerPrimitiveArgs} from '../keras_format/advanced_activation_configs';
+import {InitializerIdentifier} from '../keras_format/initializer_config';
 import {Shape} from '../keras_format/types';
 import {getRegularizer, Regularizer, serializeRegularizer} from '../regularizers';
 import {Kwargs} from '../types';
