+++ src/layers/convolutional.ts
@@ -19,12 +19,15 @@ import {Activation, getActivation, serializeActivation} from '../activations';
 import {imageDataFormat} from '../backend/common';
 import * as K from '../backend/tfjs_backend';
 import {checkDataFormat, checkPaddingMode, DataFormat, PaddingMode} from '../common';
-import {Constraint, ConstraintIdentifier, getConstraint, serializeConstraint} from '../constraints';
+import {Constraint, getConstraint, serializeConstraint} from '../constraints';
 import {InputSpec, Layer, LayerArgs, LayerNonSerializableArgs} from '../engine/topology';
 import {NotImplementedError, ValueError} from '../errors';
-import {getInitializer, Initializer, InitializerIdentifier, serializeInitializer} from '../initializers';
+import {getInitializer, Initializer, serializeInitializer} from '../initializers';
+import {ConstraintIdentifier} from '../keras_format/constraint_config';
+import {InitializerIdentifier} from '../keras_format/initializer_config';
+import {RegularizerIdentifier} from '../keras_format/regularizer_config';
 import {Shape} from '../keras_format/types';
-import {getRegularizer, Regularizer, RegularizerIdentifier, serializeRegularizer} from '../regularizers';
+import {getRegularizer, Regularizer, serializeRegularizer} from '../regularizers';
 import {Kwargs} from '../types';
 import {convOutputLength, deconvLength, normalizeArray} from '../utils/conv_utils';
 import * as generic_utils from '../utils/generic_utils';
