+++ src/layers/recurrent.ts
@@ -18,13 +18,16 @@ import {DataType, serialization, Tensor, tidy, util} from '@tensorflow/tfjs-core
 import {Activation, ActivationIdentifier, getActivation, serializeActivation} from '../activations';
 import {getScalar} from '../backend/state';
 import * as K from '../backend/tfjs_backend';
-import {Constraint, ConstraintIdentifier, getConstraint, serializeConstraint} from '../constraints';
+import {Constraint, getConstraint, serializeConstraint} from '../constraints';
 import {InputSpec, LayerNonSerializableArgs, SymbolicTensor} from '../engine/topology';
 import {Layer, LayerArgs} from '../engine/topology';
 import {AttributeError, NotImplementedError, ValueError} from '../errors';
-import {getInitializer, Initializer, InitializerIdentifier, Ones, serializeInitializer} from '../initializers';
+import {getInitializer, Initializer, Ones, serializeInitializer} from '../initializers';
+import {ConstraintIdentifier} from '../keras_format/constraint_config';
+import {InitializerIdentifier} from '../keras_format/initializer_config';
+import {RegularizerIdentifier} from '../keras_format/regularizer_config';
 import {Shape} from '../keras_format/types';
-import {getRegularizer, Regularizer, RegularizerIdentifier, serializeRegularizer} from '../regularizers';
+import {getRegularizer, Regularizer, serializeRegularizer} from '../regularizers';
 import {Kwargs, RnnStepFunction} from '../types';
 import * as math_utils from '../utils/math_utils';
 import {getExactlyOneShape, getExactlyOneTensor, isArrayOfShapes} from '../utils/types_utils';
